refactor: converte componentes locais em comp globais
