+++ src/pages/Register/components/ProgressBar/index.js
