+++ src/pages/Register/components/ProgressBar/style.js
