+++ src/pages/Register/components/species/index.js
@@ -4,7 +4,7 @@ import CatIcon from '../../../../assets/img/icon-home-cat.png';
 
 //import InfoBox from '../../../../components/InfoBox';
 import PrimaryButton from '../../../../components/PrimaryButton';
-import ProgressBar from '../../../../components/ProgressBar';
+import ProgressBar from '../ProgressBar';
 
 import {
   Container,
